% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_postcode_levels}
\alias{sim_postcode_levels}
\title{Simulate the levels and their sizes in a high-cardinality feature}
\usage{
sim_postcode_levels(nlevels = 100L, seed = 1001)
}
\arguments{
\item{nlevels}{number of levels to generate}

\item{seed}{random seed}
}
\value{
a data frame of postal codes and sizes
}
\description{
Simulate the levels and their sizes in a high-cardinality feature
}
\note{
The code is derived from the example described in the "rare levels"
vignette in the \code{vtreat} package.
}
\examples{
df_levels <- sim_postcode_levels(nlevels = 500, seed = 42)
head(df_levels)
}
