% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cud.R
\name{cud}
\alias{cud}
\title{Masataka Okabe and Kei Ito's Color Universal Design palette}
\usage{
cud(x, shift = TRUE, reverse = FALSE)
}
\arguments{
\item{x}{Vector, color index.}

\item{shift}{Start from the second color in the CUD palette?}

\item{reverse}{Reverse the color order?}
}
\value{
A vector of color hex values.
}
\description{
Masataka Okabe and Kei Ito's Color Universal Design palette
}
\examples{
barplot(rep(1, 7), col = cud(1:7))
barplot(rep(1, 8), col = cud(1:8, shift = FALSE))
barplot(rep(1, 8), col = cud(1:8, shift = FALSE, reverse = TRUE))
}
